requirements.txt
blinker==1.8.2
cachelib==0.13.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Faker==25.6.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
idna==3.7
instaloader==4.11
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
msgspec==0.18.6
mysql-connector-python==9.2.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
urllib3==2.2.1
uuid==1.30
Werkzeug==3.0.3
